% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_info.R
\name{read_file_info_}
\alias{read_file_info_}
\title{Read File Information from CSV File}
\usage{
read_file_info_(file, ...)
}
\arguments{
\item{file}{path to CSV file containing file information as stored with
\code{\link{write_file_info}}}

\item{\dots}{further arguments passed to
\code{fakin.path.app:::read_csv_fread})}
}
\description{
Read File Information from CSV File
}
\keyword{internal}
