% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny App}
\usage{
run_app(path_database = NULL, ...)
}
\arguments{
\item{path_database}{if not \code{NULL} the path to a folder containing 
text files with path information}

\item{\dots}{further \code{key = value} pairs to be used as global variables}
}
\description{
Run the Shiny App
}
